feat: implement Counter service with increment, add, and get functionalities; add module declaration
